+++ src/Compute/Compute/Models/PSVirtualMachine.cs
@@ -54,6 +54,9 @@ namespace Microsoft.Azure.Commands.Compute.Models
         // Gets or sets the property of 'Location'
         public string Location { get; set; }
 
+        // Gets or sets the property of 'EdgeZone'
+        public ExtendedLocation ExtendedLocation { get; set; }
+
         // Gets or sets the license type
         public string LicenseType { get; set; }